export utility components from a separate index
Using commonjs exports and exporting them from the main index means that if a user is importing the components into a bundler, the bundler will also try to bundle the main adapter code, which is not meant for bundling and will error out.
