+++ examples/react/components/path/path.jsx
@@ -1,6 +1,6 @@
 const React = require('react');
 
-const { usePath } = require('@frctl/react');
+const { usePath } = require('@frctl/react/components');
 
 const Path = () => {
     const path = usePath();
