+++ packages/react/components/index.js
@@ -1,0 +1,7 @@
+const PathProvider = require('./path-provider');
+
+module.exports = {
+    PathProvider: PathProvider,
+    PathContext: PathProvider.PathContext,
+    usePath: PathProvider.usePath,
+};
