+++ packages/react/package.json
@@ -3,7 +3,6 @@
   "version": "0.2.0",
   "description": "React template adapter for Fractal.",
   "main": "index.js",
-  "types": "index.d.ts",
   "repository": {
     "type": "git",
     "url": "https://github.com/frctl/fractal.git",
