--- packages/react/src/components/index.js
@@ -1,3 +1,0 @@
-module.exports = {
-    PathProvider: require('./path-provider'),
-};
